FIX: Windows build issues for new test modules
This change may not work when CIP libraries are built as DLLs. Specified
to use static libraries explicitly in CommandLineTools/CMakeLists.txt.
Fixed other Windows build issues such as fabs and time.

+++ CommandLineTools/GenerateLesionSegmentation/GenerateLesionSegmentation.cxx
@@ -133,7 +133,7 @@ int main( int argc, char * argv[] )
   InputImageType::SizeType roiSize;
   for (unsigned int i = 0; i < ImageDimension; i++)
     {
-    roiSize[i] = fabs(pi2[i] - pi1[i]);
+    roiSize[i] = fabs(double(pi2[i] - pi1[i]));
     startIndex[i] = (pi1[i]<pi2[i])?pi1[i]:pi2[i];
     }
   InputImageType::RegionType roiRegion( startIndex, roiSize );
